Flexibility changes
Added flexible constructors and descriptive comments for multi-texture,
colorized, and light-providing golems.  Creative-pick-block return now
much more customizable for these golems.

+++ com/golems/content/BlockPowerProvider.java
@@ -38,7 +38,7 @@ public class BlockPowerProvider extends BlockLightProvider
     @Override
     public int getMetaFromState(IBlockState state)
     {
-        return state.getValue(POWER);
+        return ((Integer)state.getValue(POWER)).intValue();
     }
 
     @Override
